discord-bot/pyproject.toml
[tool.poetry]
name = "afc-squad-discord-bot"
version = "0.1.0"
description = ""
authors = ["Afc_Adinho <afcadinho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
discord = "^2.3.2"
python-dotenv = "^1.1.0"
pynacl = "^1.5.0"
flask = "^3.1.0"
pillow = "^11.2.1"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
